pyproject.toml
[tool.poetry]
name = "gamification-reborn"
version = "0.1.0"
description = ""
authors = ["Mikhail Panin <mokybrow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.1"}
sqlalchemy = {extras = ["all"], version = "^2.0.20"}
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
isort = "^5.12.0"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
